:zap: "Add IdentityServer and Google authentication support"
This commit includes changes to support authentication and registration in the Questly API. IdentityServer and Google authentication configurations are added to appsettings.json. Two new classes, ApplicationDbContext and Player, are created for user management. Furthermore, necessary packages to support these changes are included in QuestlyApi.csproj. Essential middleware is added to WebApplication in Program.cs to connect the authentication flows.