Update authentication methods in AuthController
Refactored the methods in the AuthController to use JWT (JSON Web Tokens) for authentication, replaced the refresh token mechanism with directly generating JWTs. This change simplifies the authentication process and improves app's security.

The change removed the RefreshTokens and associated migration files which are no longer necessary. The AuthController's SignIn method now generates a JWT directly after successful Google Sign-In process. The RefreshToken method is replaced with Login method that authenticates users with their credentials and generates a JWT.

Also, SignUp method is introduced to enable user registration, generating a hashed password stored alongside user's email. This adds capability for users to sign up and then log in using their credentials instead of Google sign-in only.

Cleaned up some unused usings in the files and handled necessary exceptions for robustness.