Add token-refresh functionality with JWT utils
Implement functionality for secure token refresh. Also, includes the introduction of JWT utils to support authorization and JWT handling. RefreshToken entity is added to the database and DbContext along with initial migration script generated. A rename in controllers and updates in Player entity are also performed as part of improving the authentication module. AppSettings model is also updated with new settings for JWT Secret and TokenTTL. Further, added new JWT Middleware to handle token validation and claims extraction from JWT.